pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "radarange-orchestrator"  # Dash in project name
version = "0.1.0"
description = "Orchestration framework for LLM workflows"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "huggingface-hub>=0.30.1",
    "ipython>=8.35.0",
    "pydantic>=2.11.3",
    "typing-extensions>=4.0.0",
    "ipykernel>=6.22.0",
    "html2text>=2025.4.15",
    "readability-lxml>=0.3.2",
    "duckduckgo-search>=8.0.2",
    "pymupdf>=1.25.5",
    "lmstudio>=1.3.0",
]
[project.optional-dependencies]
local = ["llama-cpp-python>=0.3.9"]

[tool.setuptools.packages.find]
where = ["."]  # Search from root directory
include = ["radarange_orchestrator*"]  # Package pattern
exclude = ["examples*", "models*", "notebooks*", ".venv"]  # Exclude non-package dirs

[project.scripts]
download_model = 'models.download_model:main'

[tool.uv.pip]
no-binary = ["llama-cpp-python"]
config-settings = { "cmake.define.GGML_CUDA" = "on" }

[tool.ruff]
target-version = 'py310'
extend-include = ["*.ipynb"]
exclude = [
    '.git',
    '.ruff_cache',
    '.venv',
    'venv',
]

# Configure single quotes
[tool.ruff.format]
quote-style = 'single'